Update ecmaVersion test runner to include comments
By default, the `ecmaVersion` test runner was not asking the parser for
comments. Now, when the source filename contains `comment`, it sets
`comment: true` and includes comments in the expected result.

This extends the precedent set by filenames containing strings like
`"non-strict"`, `"edge-cases"`, `"modules"`, and a few others, all of
which put the test runner in slightly different modes.